% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{activeInput}
\alias{activeInput}
\title{Function create mock for shiny input}
\usage{
activeInput(env = new.env(), ...)
}
\arguments{
\item{env}{- enviroment that will be turn into activeInput}

\item{...}{- list of named inital values}
}
\description{
Function create mock for shiny input
}
