% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{callReaction}
\alias{callReaction}
\title{Function call callback function from list with argument message}
\usage{
callReaction(type, reactions, message)
}
\arguments{
\item{type}{- type of reaction}

\item{reactions}{- list of named callback functions}

\item{message}{- argument to call the function with}
}
\description{
Function call callback function from list with argument message
}
