% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{clean.error}
\alias{clean.error}
\title{Helper function that removes battery marker from error message}
\usage{
clean.error(cond)
}
\arguments{
\item{cond}{- condition object}
}
\description{
Helper function that removes battery marker from error message
}
