% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{clearMock}
\alias{clearMock}
\title{function restore original shiny function}
\usage{
clearMock(name, env, package = "shiny")
}
\arguments{
\item{name}{- name of the function}

\item{env}{- environment used}

\item{package}{- string with name of the pacakge}
}
\description{
function restore original shiny function
}
