% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{clearMocks}
\alias{clearMocks}
\title{Helper function to be used at the end of test files (useful if same session is used
to run test and application e.g. RStudio)}
\usage{
clearMocks()
}
\description{
Helper function to be used at the end of test files (useful if same session is used
to run test and application e.g. RStudio)
}
