% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{condition}
\alias{condition}
\title{Helper condition object}
\usage{
condition(subclass, message, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{- string vector for the subclass}

\item{message}{- message string of a given condition}

\item{call}{- for stack trace}

\item{...}{- any extra attributes that should be added to the condition}
}
\description{
Helper condition object
}
