% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{continue}
\alias{continue}
\title{function restart the evaluation of the code after the signal to stop
the propagation of event handlers in nested withHandlers}
\usage{
continue()
}
\description{
function restart the evaluation of the code after the signal to stop
the propagation of event handlers in nested withHandlers
}
