% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{create.error}
\alias{create.error}
\title{create structure that can be used to signal error in applications}
\usage{
create.error(cond, meta = NULL)
}
\arguments{
\item{cond}{- input from withCallingHandlers it should be unexpected error in app}

\item{meta}{- addition extra data that should be added into meta property}
}
\description{
create structure that can be used to signal error in applications
}
