% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{error}
\alias{error}
\title{helper function that can be used in exception handler to trigger error handler}
\usage{
error(message = NULL, bubble = FALSE)
}
\arguments{
\item{message}{- optional message that should be character string}

\item{bubble}{- if set to TRUE the message is marked so it can be tracked
while it propagates through nested handlers}
}
\description{
helper function that can be used in exception handler to trigger error handler
}
