% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{escapeRegex}
\alias{escapeRegex}
\title{function taken from Hmisc package source code (no need to whole package)}
\usage{
escapeRegex(string)
}
\arguments{
\item{string}{- string that will be escaped}
}
\description{
function taken from Hmisc package source code (no need to whole package)
}
