% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{extractActiveNames}
\alias{extractActiveNames}
\title{Traverse substitute expressions and function invocation and extract all references to active elements
created by activeInput and uiOutput}
\usage{
extractActiveNames(arg)
}
\arguments{
\item{data}{- named list: expr - substitute expr, env - parent env (value returned from renderUI)}
}
\description{
Traverse substitute expressions and function invocation and extract all references to active elements
created by activeInput and uiOutput
}
