% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{get.exceptions}
\alias{get.exceptions}
\title{helper function that returns exception handler}
\usage{
get.exceptions(session = NULL)
}
\arguments{
\item{session}{- optional shiny session, if omitted the global handlers are used}
}
\description{
helper function that returns exception handler
}
