% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{get.object}
\alias{get.object}
\title{Recursive function that get nested object and prop name for give object access exression}
\usage{
get.object(item, env)
}
\arguments{
\item{item}{- substitute expression}

\item{env}{- environment for this expression}
}
\description{
Recursive function that get nested object and prop name for give object access exression
}
