% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{getRoot}
\alias{getRoot}
\title{helper recursive function that can't be written as method}
\usage{
getRoot(node)
}
\arguments{
\item{node}{- battery component instance}
}
\description{
helper recursive function that can't be written as method
}
