% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\docType{data}
\name{global}
\alias{global}
\title{Global varialbe added to base component that hold each classes and component list
for getById method}
\format{
An object of class \code{environment} of length 6.
}
\usage{
global
}
\description{
Global varialbe added to base component that hold each classes and component list
for getById method
}
\keyword{datasets}
