% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{handle.error}
\alias{handle.error}
\title{global handler for errors that print exception if user didn't process it}
\usage{
handle.error(error, meta = NULL, session = NULL)
}
\arguments{
\item{error}{- error structure created by \code{create.error}}

\item{meta}{- context of the error (origin, type, ...)}

\item{session}{- optional shiny session used to find the exception handlers}
}
\description{
global handler for errors that print exception if user didn't process it
}
