% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{handle.exceptions}
\alias{handle.exceptions}
\title{function that invoke global exception handler based on cond data}
\usage{
handle.exceptions(cond, meta = NULL, session = NULL)
}
\arguments{
\item{cond}{- structure with classes that indicate exception}

\item{meta}{- context of the exception (origin, type, ...)}

\item{session}{- optional shiny session used to find the exception handlers}
}
\description{
function that invoke global exception handler based on cond data
}
