% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{handleErrors}
\alias{handleErrors}
\title{Helper function used by observeEvent}
\usage{
handleErrors(expr, reactionsPerType = list())
}
\arguments{
\item{expr}{- expression}

\item{reactionsPerType}{- list of named callbacks}
}
\description{
Helper function used by observeEvent
}
