% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{indent}
\alias{indent}
\title{add spaces before the string}
\usage{
indent(n, string)
}
\arguments{
\item{n}{- number of spaces}

\item{string}{- string that will be added after spaces}
}
\description{
add spaces before the string
}
