% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{invoke}
\alias{invoke}
\title{Function call function fn with the only arguments it accept}
\usage{
invoke(fn, ...)
}
\arguments{
\item{fn}{- function to be called}

\item{...}{- list of arguments}
}
\description{
Function call function fn with the only arguments it accept
}
