% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{is.active.binding}
\alias{is.active.binding}
\title{Function check if name is active biding inside environment activeInput or Output}
\usage{
is.active.binding(name, env)
}
\arguments{
\item{name}{- string}

\item{env}{- active environment to test}
}
\description{
Function check if name is active biding inside environment activeInput or Output
}
