% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{is.active.input}
\alias{is.active.input}
\title{Function for checking if object is actie input - used by extractActiveInputs}
\usage{
is.active.input(obj)
}
\description{
Function for checking if object is actie input - used by extractActiveInputs
}
