% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{is.active.output}
\alias{is.active.output}
\title{Function check if obj is enviroment that is result of activeOutput function}
\usage{
is.active.output(obj)
}
\arguments{
\item{obj}{- any object}
}
\description{
Function check if obj is enviroment that is result of activeOutput function
}
