% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{is.dbbracket.variable}
\alias{is.dbbracket.variable}
\title{Function check if argumnet is expression foo[[bar]]}
\usage{
is.dbbracket.variable(item)
}
\arguments{
\item{item}{- substitute expression}
}
\description{
Function check if argumnet is expression foo[[bar]]
}
