% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{is.dolar.variable}
\alias{is.dolar.variable}
\title{Function check if expression is foo$bar}
\usage{
is.dolar.variable(item)
}
\arguments{
\item{item}{- substitute expression}
}
\description{
Function check if expression is foo$bar
}
