% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{is.function.call}
\alias{is.function.call}
\title{function check if expression from environment is function call}
\usage{
is.function.call(expr, env)
}
\description{
function check if expression from environment is function call
}
