% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{is.inline.fucntion}
\alias{is.inline.fucntion}
\title{Function check if expression is function passed as argument}
\usage{
is.inline.fucntion(item)
}
\arguments{
\item{item}{- substitute expression}
}
\description{
Function check if expression is function passed as argument
}
