% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{is.internal.function}
\alias{is.internal.function}
\title{Function return true if function is internal - used to prevent infinite recursion}
\usage{
is.internal.function(fn.body)
}
\description{
Function return true if function is internal - used to prevent infinite recursion
}
