% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{is.silent}
\alias{is.silent}
\title{helper function that check if error that should be ignored}
\usage{
is.silent(x)
}
\arguments{
\item{x}{- condition object}
}
\description{
helper function that check if error that should be ignored
}
