% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{is.variable}
\alias{is.variable}
\title{Higher order function for creating functions for checking if item is foo$bar or foo[[bar]]}
\usage{
is.variable(chr)
}
\arguments{
\item{chr}{- character to match as first item in expression}
}
\description{
Higher order function for creating functions for checking if item is foo$bar or foo[[bar]]
}
