% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{isNullEvent}
\alias{isNullEvent}
\title{Function helper taken from \code{shiny::observeEvent} implementation
from https://github.com/rstudio/shiny/blob/master/R/reactives.R}
\usage{
isNullEvent(value)
}
\arguments{
\item{value}{- value from calling output of \code{shiny::exprToFunction}}
}
\description{
Function helper taken from \code{shiny::observeEvent} implementation
from https://github.com/rstudio/shiny/blob/master/R/reactives.R
}
