% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{isolate}
\alias{isolate}
\title{name used only inside renderUI in substitute phase}
\usage{
isolate(x)
}
\arguments{
\item{x}{- reactive value}
}
\description{
name used only inside renderUI in substitute phase
}
