% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load}
\alias{load}
\title{Function load all R file components from directory}
\usage{
load(path, pattern = "*.R$", ignore = NULL, recursive = TRUE)
}
\arguments{
\item{path}{- path to directory}

\item{pattern}{- pattern used to selected the files (default all R files)}

\item{ignore}{- string or vector of strings with fillanme files that should be ignored}

\item{recursive}{- if set to TRUE (default) loading of files will be recursive}
}
\description{
Function load all R file components from directory
}
