% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{logMessage}
\alias{logMessage}
\title{Function print message with type and optional stack trace}
\usage{
logMessage(msgData, msgType, stackTrace = NULL)
}
\arguments{
\item{msgData}{- list}

\item{msgType}{- strings}

\item{stackTrace}{- stack framees}
}
\description{
Function print message with type and optional stack trace
}
