% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{make.uiOutput}
\alias{make.uiOutput}
\title{function that return uiOutput function for given enviromnet}
\usage{
make.uiOutput(env)
}
\arguments{
\item{env}{- enviroment that should be active output}
}
\description{
function that return uiOutput function for given enviromnet
}
