% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{makeReactiveBinding}
\alias{makeReactiveBinding}
\title{Mock for shiny::makeReactiveBinding to be injected into components
limiation it can't be called after value is added to environment (it will work in components)}
\usage{
makeReactiveBinding(name, env)
}
\description{
Mock for shiny::makeReactiveBinding to be injected into components
limiation it can't be called after value is added to environment (it will work in components)
}
