% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{merge.env}
\alias{merge.env}
\title{Helper function that create single environment from environments in arguments}
\usage{
\method{merge}{env}(...)
}
\arguments{
\item{...}{- any number of  environments objects}
}
\description{
Helper function that create single environment from environments in arguments
}
