% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{merge.props}
\alias{merge.props}
\title{Merge two lists into desc list - destructive}
\usage{
\method{merge}{props}(desc, src)
}
\arguments{
\item{desc}{- detination list}

\item{src}{- source list}
}
\description{
Merge two lists into desc list - destructive
}
