% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{mock}
\alias{mock}
\title{function create single mock for shiny function}
\usage{
mock(name, mock, env, package = "shiny")
}
\arguments{
\item{name}{- string with name of the function}

\item{mock}{- function used as mock for name in package}

\item{env}{- additional environment that is also updated}

\item{package}{- name of the package}
}
\description{
function create single mock for shiny function
}
