% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{new.static.env}
\alias{new.static.env}
\title{Helper where new static default values can be added
this static env should be the same for each Component per shiny session}
\usage{
new.static.env()
}
\description{
Helper where new static default values can be added
this static env should be the same for each Component per shiny session
}
