% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{observeEventMock}
\alias{observeEventMock}
\title{Function used the same as battery::observeEvent (based on shiny::observeEvent)
that use active binding input mocks - the work almost the same as shiny::observeEvent but it
destroy previous created observer, so there are no duplicates}
\usage{
observeEventMock(value, expr, observerName = NULL, once = FALSE,
  ignoreInit = FALSE, handler.env = parent.frame(), ...)
}
\description{
Function used the same as battery::observeEvent (based on shiny::observeEvent)
that use active binding input mocks - the work almost the same as shiny::observeEvent but it
destroy previous created observer, so there are no duplicates
}
