% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observeEvent.R
\name{observerExists}
\alias{observerExists}
\title{check if handler is exists}
\usage{
observerExists(observer, handlerBody)
}
\arguments{
\item{observer}{- list returned from observeEvent}

\item{handlerBody}{- expression}
}
\description{
check if handler is exists
}
