% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{observeWrapperMock}
\alias{observeWrapperMock}
\title{Function used the same as \code{battery:::observeWrapper} (based on \code{shiny::observeEvent})}
\usage{
observeWrapperMock(
  eventExpr,
  handlerExpr,
  event.env = parent.frame(),
  handler.env = parent.frame(),
  ignoreInit = FALSE,
  ignoreNULL = TRUE,
  once = FALSE,
  ...
)
}
\arguments{
\item{eventExpr}{- same as in \code{shiny::observeEvent}}

\item{handlerExpr}{- same as in \code{shiny::observeEvent}}

\item{event.env}{- same as in \code{shiny::observeEvent}}

\item{handler.env}{- same as in \code{shiny::observeEvent}}

\item{ignoreInit}{- same as in \code{shiny::observeEvent}}

\item{ignoreNULL}{- same as in \code{shiny::observeEvent}}

\item{once}{- same as in \code{shiny::observeEvent}}

\item{...}{- reset the params from \code{battery:::observeWrapper}}
}
\value{
list with destroy method - same as \code{shiny::observeEvent}
}
\description{
Function used the same as \code{battery:::observeWrapper} (based on \code{shiny::observeEvent})
}
