% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\docType{data}
\name{originals}
\alias{originals}
\title{Env that store original functions from shiny that are ovewritten by mocks in
useMocks and later can be restored using clearMocks.}
\format{An object of class \code{environment} of length 0.}
\usage{
originals
}
\description{
Env that store original functions from shiny that are ovewritten by mocks in
useMocks and later can be restored using clearMocks.
}
\keyword{datasets}
