% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{parse.function}
\alias{parse.function}
\title{Function is parsing code like `foo() + bar()` that have type of language
and it also detect cases of `foo()`, `x$foo()` or `x[[name]]()`
We do this to find functions and methods calls and extract names
from inside of the functions bodies - body return same data
as substitute so we can call extractActiveNames on body}
\usage{
parse.function(item, env)
}
\arguments{
\item{item}{- deparse expression}

\item{env}{- enviroment}
}
\description{
Function is parsing code like `foo() + bar()` that have type of language
and it also detect cases of `foo()`, `x$foo()` or `x[[name]]()`
We do this to find functions and methods calls and extract names
from inside of the functions bodies - body return same data
as substitute so we can call extractActiveNames on body
}
