% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{pause}
\alias{pause}
\title{helper function that can be used in exception handler to pause just this context}
\usage{
pause()
}
\description{
helper function that can be used in exception handler to pause just this context
}
