% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{r6.class.add}
\alias{r6.class.add}
\title{helper function for adding properties to R6Class}
\usage{
r6.class.add(class, seq, spy = FALSE)
}
\description{
helper function for adding properties to R6Class
}
