% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{recur.list}
\alias{recur.list}
\title{Helper function used for debugging it convert nested expression structure tree to list}
\usage{
recur.list(expr)
}
\arguments{
\item{expr}{- substitute expression}
}
\description{
Helper function used for debugging it convert nested expression structure tree to list
}
