% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/mocks.R
\name{renderUI}
\alias{renderUI}
\title{renderUI wrapper with force argument (all params are the )}
\usage{
renderUI(expr)

renderUI(expr)
}
\arguments{
\item{expr}{- same as in shiny::renderUI}

\item{env}{- same as in shiny::renderUI}

\item{quoted}{- same as in shiny::renderUI}

\item{outputArgs}{- same as in shiny::renderUI}

\item{force}{- it will always call renderUI on init, even if shiny think that it don't need to}

\item{expr}{- any expression}
}
\description{
renderUI wrapper with force argument (all params are the )

RenderUI just send exression to output active prop, the prop need to be added first
if renderUI is called in constructor and it use self$ns() you can pass component.id to
constructor so instance will have same id and you can generate the name before constructor
is called - using: output$new(name)
the output will parse the expression and bind input and output
note: input can have different name for instance events in components
}
