% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{report.error}
\alias{report.error}
\title{helper function that print the context in which the error was thrown}
\usage{
report.error(error, meta = NULL)
}
\arguments{
\item{error}{- error structure created by \code{create.error}}

\item{meta}{- context of the error (origin, type, ...)}
}
\description{
helper function that print the context in which the error was thrown
}
