% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{reset.exceptions}
\alias{reset.exceptions}
\title{helper function that reset exceptions}
\usage{
reset.exceptions(session = NULL)
}
\arguments{
\item{session}{- optional shiny session, if omitted the global handlers are reset}
}
\description{
helper function that reset exceptions
}
