% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{resolve.exceptions}
\alias{resolve.exceptions}
\title{helper function that return exception handlers that apply to given session}
\usage{
resolve.exceptions(session = NULL)
}
\arguments{
\item{session}{- optional shiny session}
}
\description{
handlers registered for the session take precedence, if the session has no
handlers it falls back to the global ones
}
