% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{safe.get}
\alias{safe.get}
\title{function saftly return name from enviroment(s) and don't throw error
if name don't exists}
\usage{
safe.get(name, env)
}
\arguments{
\item{name}{- string}

\item{env}{- environment or list of environments}
}
\description{
function saftly return name from enviroment(s) and don't throw error
if name don't exists
}
