% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{Session}
\alias{Session}
\title{Shiny Session Mock class}
\description{
Mock for Session object, only some tests require this right now
but more tests may require this session object in the future
Right now only services need session$destroy() method to clear
services so you can create same service in more then one test

with version 0.5.0 session mock can be used as the only object
passed to root component constructor.
}
\examples{

TestComponent <- battery::component(
  "TestComponent",
  public = list(
    constructor = function() {
       print(self$session$token)
    }
  )
)
session <- battery::Session$new("Test")
input <- activeInput()
output <- activeOutput()
component <- TestComponent$new(input = input, output = output, session = session)
}
\keyword{mock}
\keyword{testing}
\keyword{unittests}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{token}}{- string used by battery to distinguish users}

\item{\code{input}}{- shiny input}

\item{\code{output}}{- shiny output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Session-new}{\code{Session$new()}}
\item \href{#method-Session-destroy}{\code{Session$destroy()}}
\item \href{#method-Session-onSessionEnded}{\code{Session$onSessionEnded()}}
\item \href{#method-Session-clone}{\code{Session$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-new"></a>}}
\if{latex}{\out{\hypertarget{method-Session-new}{}}}
\subsection{Method \code{new()}}{
Session mock constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$new(
  token = NULL,
  input = battery::activeInput(),
  output = battery::activeOutput()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{- optional token used for testing to create different
users that should get different data for services and globals}

\item{\code{input}}{- mock for shiny input}

\item{\code{output}}{- mock for shiny output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-destroy"></a>}}
\if{latex}{\out{\hypertarget{method-Session-destroy}{}}}
\subsection{Method \code{destroy()}}{
Mock for destroy session

it will trigger handlers added by \code{onSessionEnded}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-onSessionEnded"></a>}}
\if{latex}{\out{\hypertarget{method-Session-onSessionEnded}{}}}
\subsection{Method \code{onSessionEnded()}}{
Mock for the function that add handler on session destroy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$onSessionEnded(fn)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{- function that will be called when \code{destroy}
is called}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Session-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Session-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
