% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{set.exceptions}
\alias{set.exceptions}
\title{helper function that set exception uncoditionaly, old handler is discarded}
\usage{
set.exceptions(handler, session = NULL)
}
\arguments{
\item{handler}{- named list of exception handlers}

\item{session}{- optional shiny session, if omitted the global handlers are set}
}
\description{
helper function that set exception uncoditionaly, old handler is discarded
}
