% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{set.frame}
\alias{set.frame}
\title{Function overwrite value in parent frame env
It start searching for variable starting from frame parameter}
\usage{
set.frame(value, name = NULL, frame = 1)
}
\description{
Function overwrite value in parent frame env
It start searching for variable starting from frame parameter
}
