% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{signal}
\alias{signal}
\title{signal exception in applications}
\usage{
signal(class, message, ...)
}
\arguments{
\item{class}{- string vector that indicate class of the exception}

\item{message}{- string that indicate given exception}

\item{...}{- eny extra data that should be added to the exception}
}
\description{
signal exception in applications
}
