% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{stop}
\alias{stop}
\title{helper function that can be used in exception handler to stop whole application}
\usage{
stop()
}
\description{
helper function that can be used in exception handler to stop whole application
}
