% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{activeOutput}
\alias{activeOutput}
\title{Function create active binding output mock to be used with renderUI mock}
\usage{
activeOutput(...)
}
\arguments{
\item{...}{- named initial active inputs}
}
\description{
Function create active binding output mock to be used with renderUI mock
}
