% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{clean}
\alias{clean}
\title{helper function that clean message and remove classes from cond object}
\usage{
clean(cond)
}
\description{
helper function that clean message and remove classes from cond object
}
