% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\docType{data}
\name{Component}
\alias{Component}
\title{Base class for components}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Component
}
\arguments{
\item{event}{- name of internal event or input id}

\item{handler}{- function that should have value and target parameters}

\item{input}{- boolean that's indicate if event should be added to input}

\item{enabled}{- boolean that enable event to easy toggle event}

\item{init}{- indicate if event should be triggered on init}
}
\description{
Base class for components
}
\keyword{datasets}
