% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{component}
\alias{component}
\title{Helper function for defining components with additional static field. It can also be used
so you don't confuse battery component with normal R6Class}
\usage{
component(classname, public = NULL, private = NULL, static = NULL,
  inherit = battery::Component, spy = FALSE, ...)
}
\arguments{
\item{public}{- list of public functions and properties}

\item{private}{- list of private functions and properties}

\item{static}{- list of fields that will stay the same for every instance of the component}

\item{inherit}{- base class - if not specifed it will inherit from Base class (battery::Component)}
}
\description{
Helper function for defining components with additional static field. It can also be used
so you don't confuse battery component with normal R6Class
}
