% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{error}
\alias{error}
\title{helper function that can be used in exception handler to trigger error handler}
\usage{
error(message = NULL)
}
\arguments{
\item{message}{- optional message that should be character string}
}
\description{
helper function that can be used in exception handler to trigger error handler
}
