% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventEmitter.R
\name{EventEmitter}
\alias{EventEmitter}
\title{Event emitter is inspired by multiple implementation in JavaScript
it can be used indepentely of battery events to send data from one component
to different one, without the need to broadcast and emit events
it can be used with services, in addition to can also be used as reactive
values in shiny reactive context like renderUI or battery component render
function}
\description{
Event emitter is inspired by multiple implementation in JavaScript
it can be used indepentely of battery events to send data from one component
to different one, without the need to broadcast and emit events
it can be used with services, in addition to can also be used as reactive
values in shiny reactive context like renderUI or battery component render
function

Event emitter is inspired by multiple implementation in JavaScript
it can be used indepentely of battery events to send data from one component
to different one, without the need to broadcast and emit events
it can be used with services, in addition to can also be used as reactive
values in shiny reactive context like renderUI or battery component render
function
}
\examples{

e <- EventEmitter$new()
e$on("sessionCreated", function(value, emitter) {
  print(value$name)
})

## and in different part of the application

e$emit("sessionCreated", list(name = "My Session"))

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EventEmitter$new()}}
\item \href{#method-on}{\code{EventEmitter$on()}}
\item \href{#method-emit}{\code{EventEmitter$emit()}}
\item \href{#method-off}{\code{EventEmitter$off()}}
\item \href{#method-finalize}{\code{EventEmitter$finalize()}}
\item \href{#method-clone}{\code{EventEmitter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$new(spy = FALSE, shiny = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-on"></a>}}
\subsection{Method \code{on()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$on(events, handler, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-emit"></a>}}
\subsection{Method \code{emit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$emit(name, data = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-off"></a>}}
\subsection{Method \code{off()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$off(events, handler = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventEmitter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
