% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventEmitter.R
\docType{data}
\name{EventEmitter}
\alias{EventEmitter}
\title{Event emitter is inspired by multiple implementation in JavaScript
it can be used indepentely of battery events to send data from one component
to different one, without the need to broadcast and emit events
it can be used with services, in addition to can also be used as reactive
values in shiny reactive context like renderUI or battery component render
function}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EventEmitter
}
\description{
Event emitter is inspired by multiple implementation in JavaScript
it can be used indepentely of battery events to send data from one component
to different one, without the need to broadcast and emit events
it can be used with services, in addition to can also be used as reactive
values in shiny reactive context like renderUI or battery component render
function
}
\examples{
e <- EventEmitter$new()
e$on("sessionCreated", function(value, emitter) {
  self$updateSessionComponent(value$name)
})

and in different part of the application

e$emit("sessionCreated", list(name = "My Session"))

}
\keyword{datasets}
