% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{exceptions}
\alias{exceptions}
\title{function allows to add global exception handler for all battery components}
\usage{
exceptions(handler = NULL, reset = FALSE)
}
\arguments{
\item{handler}{- list of handlers}

\item{rest}{- indicate if old handlers should be removed}
}
\description{
function allows to add global exception handler for all battery components
}
