% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{extractActiveInputs}
\alias{extractActiveInputs}
\title{Traverse substitute expression and extract all references to active elements
created by activeInput}
\usage{
extractActiveInputs(data)
}
\arguments{
\item{s}{- named list: expr - substitute expr, env - parent env (value returned from renderUI)}
}
\description{
Traverse substitute expression and extract all references to active elements
created by activeInput
}
