% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{handle.error}
\alias{handle.error}
\title{global handler for errors that print exception if user didn't process it}
\usage{
handle.error(error, finally = NULL)
}
\description{
global handler for errors that print exception if user didn't process it
}
