% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{handle.exceptions}
\alias{handle.exceptions}
\title{function that invoke global exception handler based on cond data}
\usage{
handle.exceptions(cond, finally = NULL)
}
\arguments{
\item{cond}{- structure with classes that indicate exception}
}
\description{
function that invoke global exception handler based on cond data
}
