% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{is.battery.error}
\alias{is.battery.error}
\title{helper function that check if error was triggered by battery::error function}
\usage{
is.battery.error(cond)
}
\description{
helper function that check if error was triggered by battery::error function
}
