% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load}
\alias{load}
\title{Function load all R file components from directory, component classes will be global}
\usage{
load(
  path,
  pattern = "*.R$",
  ignore = NULL,
  recursive = TRUE,
  ignore.case = TRUE,
  local = TRUE
)
}
\arguments{
\item{path}{- path to directory}

\item{pattern}{- pattern used to selected the files (default all R files)}

\item{ignore}{- string or vector of strings with fillanme files that should be ignored}

\item{recursive}{- if set to TRUE (default) loading of files will be recursive}

\item{ignore.case}{- should it ignore case of pattern when searching for files}

\item{local}{- TRUE, FALSE or enviroment use in source - if value is TRUE it will use parent.frame()}
}
\description{
Function load all R file components from directory, component classes will be global
}
