% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{make.extend}
\alias{make.extend}
\title{higher order function for creating extend static method on every battery::Component}
\usage{
make.extend(class)
}
\arguments{
\item{class}{- battery component constructor}
}
\value{
function
}
\description{
higher order function for creating extend static method on every battery::Component
}
