% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{new.global.env}
\alias{new.global.env}
\title{Global env generator function - it's added to each component one per root (component without parent)
This env is used so you can share services for one tree of components. You can't send data between
Two trees of components.}
\usage{
new.global.env()
}
\description{
Global env generator function - it's added to each component one per root (component without parent)
This env is used so you can share services for one tree of components. You can't send data between
Two trees of components.
}
