% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{observeEvent}
\alias{observeEvent}
\title{Function used same as shiny observeEvent that use active binding input mocks}
\usage{
observeEvent(value, expr = NULL)
}
\description{
Function used same as shiny observeEvent that use active binding input mocks
}
