% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{observeEventMock}
\alias{observeEventMock}
\title{Function used the same as battery::observeEvent (based on shiny::observeEvent)
that use active binding input mocks - the work almost the same as shiny::observeEvent but it
destroy previous created observer, so there are no duplicates}
\usage{
observeEventMock(eventExpr, handlerExpr, handler.env = parent.frame(),
  ignoreInit = FALSE, ignoreNULL = TRUE, observerName = NULL,
  once = FALSE, ...)
}
\arguments{
\item{eventExpr}{- same as in shiny::observeEvent}

\item{handlerExpr}{- same as in shiny::observeEvent}

\item{handler.env}{- same as in shiny::observeEvent}

\item{ignoreInit}{- same as in shiny::observeEvent}

\item{ignoreNULL}{- same as in shiny::observeEvent}

\item{observerName}{- name that will distinguish observers with same code}
}
\description{
Function used the same as battery::observeEvent (based on shiny::observeEvent)
that use active binding input mocks - the work almost the same as shiny::observeEvent but it
destroy previous created observer, so there are no duplicates
}
