% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{r6.class.add}
\alias{r6.class.add}
\title{Helper function for adding properties to R6Class}
\usage{
r6.class.add(class, seq)
}
\arguments{
\item{class}{- battery component constructor}

\item{seq}{- named list of properties and functions methods}
}
\description{
Helper function for adding properties to R6Class
}
