% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{Session}
\alias{Session}
\title{Mock for Session object, only some tests require this right now
but more tests may require this session object in the future
Right now only services need session$destroy() method to clear
services so you can create same service in more then one test}
\description{
Mock for Session object, only some tests require this right now
but more tests may require this session object in the future
Right now only services need session$destroy() method to clear
services so you can create same service in more then one test

Mock for Session object, only some tests require this right now
but more tests may require this session object in the future
Right now only services need session$destroy() method to clear
services so you can create same service in more then one test
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Session$new()}}
\item \href{#method-destroy}{\code{Session$destroy()}}
\item \href{#method-onSessionEnded}{\code{Session$onSessionEnded()}}
\item \href{#method-clone}{\code{Session$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$new(token = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-destroy"></a>}}
\subsection{Method \code{destroy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$destroy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-onSessionEnded"></a>}}
\subsection{Method \code{onSessionEnded()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$onSessionEnded(fn)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Session$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
