% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\docType{data}
\name{Session}
\alias{Session}
\title{Mock for Session object, only some tests require this right now
but more tests may require this session object in the future
Right now only services need session$destroy() method to clear
services so you can create same service in more then one test}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Session
}
\description{
Mock for Session object, only some tests require this right now
but more tests may require this session object in the future
Right now only services need session$destroy() method to clear
services so you can create same service in more then one test
}
\keyword{datasets}
