% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{set.exceptions}
\alias{set.exceptions}
\title{helper function that set exception uncoditionaly, old handler is discarded}
\usage{
set.exceptions(handler, session = NULL)
}
\description{
helper function that set exception uncoditionaly, old handler is discarded
}
