% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocks.R
\name{useMocks}
\alias{useMocks}
\title{Helper function to be used in test files that overwrite shiny functions}
\usage{
useMocks()
}
\description{
Helper function to be used in test files that overwrite shiny functions
}
