% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{withExceptions}
\alias{withExceptions}
\title{global exception handler that is used in battery instead of tryCatch}
\usage{
withExceptions(expr, error = NULL, finally = NULL)
}
\arguments{
\item{expr}{- any expression}

\item{error}{- function that will be triggered on error default NULL
it should return add add meta data create.error(cond, list(...))}

\item{finally}{- function that is always executed after exception is handled}
}
\description{
global exception handler that is used in battery instead of tryCatch
}
