% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptions.R
\name{withExceptions}
\alias{withExceptions}
\title{global exception handler that is used in battery instead of tryCatch}
\usage{
withExceptions(expr, error = NULL, finally = NULL, meta = NULL, session = NULL)
}
\arguments{
\item{expr}{- any expression}

\item{error}{- function that will be triggered on error default NULL
if added it should return add add meta data create.error(cond, list(...))
it is used internally by battery, it can safely ignored.}

\item{finally}{- function that is always executed after exception is handled}

\item{session}{- optional shiny session to create exception handler only for given session}
}
\description{
global exception handler that is used in battery instead of tryCatch
}
